+++ santashop-core/src/lib/pipes/index.ts
@@ -1,2 +1,3 @@
 export * from './nice-form-error.pipe';
+export * from './event-date.pipe';
 export * from './time-slot.pipe';